requirements.txt
streamlit
openai
yt-dlp
openai-whisper
torch
